Place the tests in the right test case
